% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQtl.R
\name{applyQtl}
\alias{applyQtl}
\title{Filter a Tibble to Obtain Values Outside a QTL}
\usage{
applyQtl(data, var, lower = NULL, upper = NULL)
}
\arguments{
\item{data}{the input tibble}

\item{var}{the name of the column on which to filter.  Uses NSE.}

\item{lower}{The lower limit.  May be \code{NULL}.}

\item{upper}{The upper limit.  May be \code{NULL}.}
}
\value{
the filtered tibble
}
\description{
Given an input tibble and a column, filter the tibble to retain only those
rows in which the value of the column lies outside the range given by
upper and lower limits, both of which are optional.
}
