% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berrySubject}
\alias{berrySubject}
\title{#' Data derived and adapted from Berry et al (2011) pp 52-63}
\format{
A tibble with 150 rows and 3 columns
\describe{
\item{Site}{The site ID}
\item{SubjectID}{The subject ID}
\item{Event}{Whether or not the subject experienced an event}
}
}
\usage{
berrySubject
}
\description{
A dataset of subject and event flags
}
\keyword{datasets}
