% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berrySummary}
\alias{berrySummary}
\title{Data adapted from Berry et al (2011) pp 52-63}
\format{
A tibble with 9 rows and 4 columns
\describe{
\item{Site}{The site ID}
\item{Subjects}{The number of subjects at the site}
\item{Events}{The number of subjects ith at least one event at the site}
\item{ObservedResponse}{The observed response rate \code{Events}/\code{Subjects} at
the site}
}
}
\usage{
berrySummary
}
\description{
A dataset of subject and event counts by site
}
\keyword{datasets}
