% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cavalryDeaths}
\alias{cavalryDeaths}
\title{The Bortkiewicz cavalry dataset}
\format{
A tibble with 280 rows and 3 variables:
\describe{
\item{Corps}{The corps}
\item{Year}{The year}
\item{Deaths}{The number of deaths in the given corps in the given year}
...
}
}
\source{
\url{https://archive.org/details/dasgesetzderklei00bortrich}
}
\usage{
cavalryDeaths
}
\description{
A dataset number of Prussian cavalry soldiers kicked to death by their horses
}
\keyword{datasets}
