% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObservedMinusExpectedPlot.R
\name{createObservedMinusExpectedPlot}
\alias{createObservedMinusExpectedPlot}
\title{Create an Observed Minus Expected Plot}
\usage{
createObservedMinusExpectedPlot(
  data,
  indexVar = SubjectIndex,
  upperActionLimit = UpperActionLimit,
  lowerActionLimit = LowerActionLimit,
  upperWarningLimit = UpperWarningLimit,
  lowerWarningLimit = LowerWarningLimit,
  warningQuantiles = c(lower = 0.01, upper = 0.99),
  permittedRates = c(lower = 0.05, upper = 0.15)
)
}
\arguments{
\item{data}{The input \code{data.frame}}

\item{indexVar}{The column in \code{data} that defines the sequence in which
participants enrolled.  Should be an intger vector with values from \code{1} to \code{n},
where \code{n} is the number of participants enrolled.}

\item{upperActionLimit}{The column in \code{data} that defines the upper action limit.
If \code{NULL}, the upper action limit is omitted from the plot.}

\item{lowerActionLimit}{The column in \code{data} that defines the lower action limit.
If \code{NULL}, the lower action limit is omitted from the plot.}

\item{upperWarningLimit}{The column in \code{data} that defines the upper warning limit.
If \code{NULL}, the upper warning limit is omitted from the plot.}

\item{lowerWarningLimit}{The column in \code{data} that defines the lower warning limit.
If \code{NULL}, the lower warning limit is omitted from the plot.}

\item{warningQuantiles}{A named numeric vector with elements \code{upper} and \code{lower}
containing the quantiles of the binomial distribution that define the warning limits.
Used only for labeling.  If \code{NULL}, the corresponding footer is omitted
from the plot.  Individual elements may be \code{NA}.}

\item{permittedRates}{A named numeric vector with elements \code{upper} and \code{lower}
containing the observed event rates that define the action limits.
Used only for labeling.  If \code{NULL}, the corresponding footer is omitted
from the plot.  Individual elements may be \code{NA}.}
}
\description{
Given an input \code{data.frame}, create the corresponding observed-expected plot with
option upper and lower warning limits and optional upper and lower action
limits
}
