% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObservedMinusExpectedTable.R
\name{createObservedMinusExpectedTable}
\alias{createObservedMinusExpectedTable}
\title{Create a data.frame containing the results of an Observed - Expected Analysis}
\usage{
createObservedMinusExpectedTable(
  data,
  timeVar,
  eventVar,
  eventArray,
  expectedRate,
  cumulativeVar = CumulativeEvents,
  observedMinusExpectedVar = ObservedMinusExpected,
  indexVar = SubjectIndex,
  maxTrialSize = NA,
  upperActionLimit = UpperActionLimit,
  lowerActionLimit = LowerActionLimit,
  upperWarningLimit = UpperWarningLimit,
  lowerWarningLimit = LowerWarningLimit,
  warningQuantiles = c(lower = 0.01, upper = 0.99),
  permittedRates = c(lower = 0.05, upper = 0.15)
)
}
\arguments{
\item{data}{The input \code{data.frame}.  Required.}

\item{timeVar}{The column in \code{data} that defines the order in which
participants enrolled. Required.}

\item{eventVar}{The column in \code{data} that defines the whether or not a
participant experienced an event. Required.}

\item{eventArray}{An array defining the values of \code{eventVar} that indicate
that the corresponding participant experienced an event.  Required.}

\item{expectedRate}{A scalar that defines the participant-level event rate
that is expected.  Required.}

\item{cumulativeVar}{The name of the column in the output \code{data.frame} that
contains the cumulative sum of events.
Uses tidy evaluation.}

\item{observedMinusExpectedVar}{The name of the column in the output \code{data.frame} that
contains the calculated observed - expected statistic for each participant.
Uses tidy evaluation.}

\item{indexVar}{The name of the column in the output \code{data.frame} that
contains the index of the participant's enrollment: an integer from \code{1} to \code{n}
where \code{n} is number of participants enrolled. Uses tidy evaluation.}

\item{maxTrialSize}{A scalar integer defining the maximum possible size of
the trial.  Can be \code{NULL}.}

\item{upperActionLimit}{The name of the column in the output \code{data.frame} that
contains the upper action limit for this participant.  Uses tidy evaluation.}

\item{lowerActionLimit}{The name of the column in the output \code{data.frame} that
contains the lower action limit for this participant.  Uses tidy evaluation.}

\item{upperWarningLimit}{The name of the column in the output \code{data.frame} that
contains the upper warning limit for this participant.  Uses tidy evaluation.}

\item{lowerWarningLimit}{The name of the column in the output \code{data.frame} that
contains the lower warning limit for this participant.  Uses tidy evaluation.}

\item{warningQuantiles}{a named numeric vector of length 2 with elements
\code{lower} and \code{upper} that define the quantiles of the binomial distribution
that define the warning limits.  Individual elements may be \code{NA}.  If \code{NULL},
\code{lowerWarningLimit} and \code{upperWarningLimit} are ignored and the corresponding
columns are not added to the output \code{data.frame}}

\item{permittedRates}{a named numeric vector of length 2 with elements
\code{lower} and \code{upper} that define the observed event rates that define the
action limits.  Individual elements may be \code{NA}.  If \code{NULL},
\code{lowerActionLimit} and \code{upperActionLimit} are ignored and the corresponding
columns are not added to the output \code{data.frame}}
}
\description{
Create a data.frame containing the results of an Observed - Expected Analysis
}
