% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObservedOverExpectedPlot.R
\name{createObservedOverExpectedPlot}
\alias{createObservedOverExpectedPlot}
\title{Create an Observed Over Expected Plot}
\usage{
createObservedOverExpectedPlot(
  data,
  n = NObserved,
  observedRate = NULL,
  alpha = 0.05,
  sides = c("two", "lower", "upper")
)
}
\arguments{
\item{data}{The tibble containing the calculated limits}

\item{n}{The column in \code{data} that defines the sample size to which the
corresponding limit(s) relate(s)}

\item{observedRate}{The column in \code{data} that defines the observed rate.
May be \code{NULL}.}

\item{alpha}{The Type 1 error rate associated with the calculated limits.
Default 0.05}

\item{sides}{the sidedness of \code{alpha}.  Either \code{two}, \code{upper} or \code{lower}.
Default \code{two}.}
}
\value{
a \code{ggplot2} object containing the requested plot
}
\description{
Given a tibble created by \code{\link{createObservedOverExpectedTable}} create an
observed over expected plot.
}
