% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createObservedOverExpectedTable.R
\name{createObservedOverExpectedTable}
\alias{createObservedOverExpectedTable}
\title{Create an Observed Over Expected Grid

Using the work of Katz et al (1978) calculate acceptable limits for the
ratio of two binomial proportions.  The Type 1 error rate can be specified,
and the limits can be one- or two-sided.}
\usage{
createObservedOverExpectedTable(
  observedData = NULL,
  observedRate = NULL,
  n = NObserved,
  nHistorical,
  nObservedRange = seq(50, 1500, 25),
  expectedRate,
  historicalRate,
  alpha = 0.05,
  sides = c("two", "lower", "upper")
)
}
\arguments{
\item{observedData}{a tibble containing the observed data that will be
compared to the calculated limit(s).  May be \code{NULL}.}

\item{observedRate}{The column in \code{observedData} that contains the observed
rates.  Ignored if \code{observedData} is NULL.  Uses tidy evaluation.}

\item{n}{The column in both \code{observedData} (if not \code{NULL}) and the output
tibble that defines the sample size to which the corresponding limit(s)
relate(s).}

\item{nHistorical}{the number of observations on which the historical rate
is based}

\item{nObservedRange}{a sequence of values for which limits should be
calculated}

\item{expectedRate}{the event rate expected in the current study.  Usually
the same as \code{historicalRate}, but need not be so.}

\item{historicalRate}{the event rate observed in the historical data}

\item{alpha}{The Type 1 error rate associated with the calculated limits.
Default 0.05}

\item{sides}{the sidedness of \code{alpha}.  Either \code{two}, \code{upper} or \code{lower}.
Default \code{two}.}
}
\value{
a tibble
}
\description{
Create an Observed Over Expected Grid

Using the work of Katz et al (1978) calculate acceptable limits for the
ratio of two binomial proportions.  The Type 1 error rate can be specified,
and the limits can be one- or two-sided.
}
