% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQtlBubblePlot.R
\name{createQtlBubblePlot}
\alias{createQtlBubblePlot}
\title{Create a QTL bubble plot}
\usage{
createQtlBubblePlot(
  d,
  x,
  y,
  size,
  group = NULL,
  limits = NULL,
  summarise = FALSE,
  summarize = summarise,
  stat = NULL,
  jitterHeight = 0,
  jitterWidth = NULL,
  showBox = TRUE,
  boxFill = "steelblue2",
  boxColour = boxFill,
  boxColor = boxColour,
  boxAlpha = 0.25,
  boxWidth = 0.5,
  showLegend = TRUE,
  ...
)
}
\arguments{
\item{d}{the data frame containing the data to be plotted}

\item{x}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> the column in \code{data} that should appear on the x-axis}

\item{y}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> the column in \code{data} that should appear on the y-axis}

\item{size}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> the column in \code{data} that defines the size of the bubbles}

\item{group}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> Default: \code{NULL} the column in \code{data} that defines the fill colour of the bubbles}

\item{limits}{Default: \code{NULL} a \code{list} defining the characteristics of the horizontal reference lines that should appear on the plot.  See Usage notes below.}

\item{summarise}{Default: \code{FALSE} Should the input data frame be summarised to derive the plotted statistics?  See Usage notes below}

\item{summarize}{Default: \code{summarise}  A US English equivalent to \code{summarise}}

\item{stat}{Default: \code{NULL}  The summary function to be used to summarise the data.  Cannot be \code{NULL} if \code{summarise} is \code{TRUE}.  See Usage notes below}

\item{jitterHeight}{Default: \code{0}.  Value of the \code{height} parameter in the call to \code{geom_jitter}.  The default value implies no vertical jittering.}

\item{jitterWidth}{Default: \code{NULL}.  Value of the \code{width} parameter in the call to \code{geom_jitter}.  The default value implies a 40\% horizontal jitter.}

\item{showBox}{Default: \code{TRUE} show an unjittered box-and-whisker plot at each value of \code{x}?.}

\item{boxFill}{Default: \code{"steelblue2"}: the colour with which the box-and-whisker plot is filled}

\item{boxColour}{Default: \code{boxFill}: the colour with which the outline of the box-and-whisker plot is drawn}

\item{boxColor}{Default: \code{boxColour}: US-English version of \code{boxColour}}

\item{boxAlpha}{Default: \code{0.25} The transparency of the box-and-whisker plot.  If \code{alpha=0}, the box is invisible.  If \code{alpha=1}, the box is totally opaque.}

\item{boxWidth}{Default: \code{0.5} The width of each box in the box-and-whisker plot, in the same units as those used on the x-axis}

\item{showLegend}{Default: \code{TRUE}  Should the legend defined by \code{group} appear in the plot}

\item{...}{additional parameters passed to \code{ggplot2} functions and \code{stat}}
}
\value{
a \code{gg} object
}
\description{
Create a QTL bubble plot
}
\section{Usage notes}{

\code{limits} should be a list of lists, each element of which defines the label (default \code{""}), \code{colour} (default \code{"black"}), line \code{type} (default \code{"solid"}), \code{y} co-ordinate (default \code{1}), \code{x} co-ordinate (default \code{0}), and vertical offset (\code{vjust}) (default \code{-1}) of a reference line.  If any element of the component lists is missing, a warning is issued and a default value supplied.  For example:  \code{limits=list(list(label="QTL (12\%)", colour="red", type="dashed", y=12, x=1.5, vjust=-1),list(label="Sec Lim (8\%)", colour="goldenrod", type="dotted", y=58, x=1.25, vjust=-1))}

If \code{summarise} is \code{TRUE}, the (participant-level) information in \code{data} is summarised before the plot is created.  \code{data} is grouped by \code{group} and \code{x} and then \code{y=stat(y)} and \code{size=sum(size)} are calculated, where \code{x}, \code{y}, \code{size} and \code{stat} are the parameters defined above.
}

