% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQtlPlot.R
\name{createQtlPlot}
\alias{createQtlPlot}
\title{Summary Plot of Observed Event Rates/Proportions}
\usage{
createQtlPlot(
  data,
  qtl = NULL,
  siteData = NULL,
  targetRange = NULL,
  observedMetric = NULL
)
}
\description{
Summary Plot of Observed Event Rates/Proportions
}
