% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.assertColumnExists}
\alias{.assertColumnExists}
\title{Throw an exception of the given column DOES NOT exist in the given data.frame}
\usage{
.assertColumnExists(data, col)
}
\arguments{
\item{data}{the data.frame to be searched}

\item{col}{the name of the column to be searched}
}
\value{
\code{NULL}
}
\description{
Throw an exception of the given column DOES NOT exist in the given data.frame
}
\examples{
# tibble::tibble(good=1) \%>\% .assertColumnExists(good) # No exception
# tibble::tibble(good=1) \%>\% .assertColumnExists(bad) # Exception
}
