% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.columnExists}
\alias{.columnExists}
\title{Determine if a column, passed using NSE, exists in a data.frame}
\usage{
.columnExists(data, col)
}
\arguments{
\item{data}{the data.frame to be searched}

\item{col}{the name of the column to be searched}
}
\value{
\code{TRUE} if the column exists, \code{FALSE} otherwise
}
\description{
Determine if a column, passed using NSE, exists in a data.frame
}
\examples{
# mtcars \%>\% .columnExists(mpg)
# mtcars \%>\% .columnExists(bad)
}
