% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBinomialInit.R
\name{.createBinomialInit}
\alias{.createBinomialInit}
\title{Create a JAGS inits suitable for use with run.jags and autorun.jags}
\usage{
.createBinomialInit(
  gammaA = c(shape = 1, scale = 10),
  gammaB = c(shape = 1, scale = 10),
  seed = NULL,
  quantiles = NULL,
  rng = c("base::Mersenne-Twister", "base::Wichmann-Hill", "base::Marsaglia-Multicarry",
    "base::Super-Duper")
)
}
\arguments{
\item{gammaA}{a vector with elements named shape and scale that define the
hyper prior for \code{a} in the Beta prior for \code{p}.}

\item{gammaB}{a vector with elements named shape and scale that define the
hyper prior for \code{b} in the Beta prior for \code{p}.}

\item{seed}{a scalar containing the seed to be used for the init.  \code{NULL},
the default, indicates that a random  seed should be used.  It
is not clear from the runjags or JAGS manuals what the range of permissible
seed values is, but positive integers appear to be acceptable.}

\item{quantiles}{Default \code{NULL}. If \code{NULL}, starting values of \code{a} and
\code{b} are chosen at random.  Otherwise, starting values are defined by the
values supplied in \code{quantiles}. See Usage Notes below.}

\item{rng}{the random number generator to be used for each init}
}
\value{
The init: a list with elements named \code{p}, \code{a}, \code{b}, \code{RNG.name} and
\code{RNG.seed}.
}
\description{
Initial values of p are chosen uniformly at random in the range 0.001 to
0.999 inclusive.
}
\section{Usage Notes}{

The \code{quantiles} parameter provides a flexible way of defining starting values
for the hyperparameters \code{a} and \code{b}.  If \code{NULL}, the default, starting values
are random values from either \code{gammaA} (for \code{a}) or \code{gammaB} (for \code{b}).
Otherwise, \code{quantiles} should be a numeric vector of length 2 with values
(named \code{a} and \code{b}) that are both either \code{NA} or in the range (0, 1).  If
\code{NA}, the corresponding staring value is chosen at random, as above.  Otherwise,
if the value is \code{x}, the starting value of the hyperparameter is the \code{x}th
quantile of the corresponding Gamma distribution.

For example, if \code{quantiles=c("a"=NA, "b"=50)}, then the starting value for \code{a}
is a random value drawn from the distribution defined by \code{gammaA} and the
starting value of \code{b} is the 50th centile of the distribution defined by
\code{gammaB}.
}

