% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.ensureLimitsAreNamed}
\alias{.ensureLimitsAreNamed}
\title{Ensures that a vector or scalar is named according to standard rules}
\usage{
.ensureLimitsAreNamed(x, decreasing = FALSE, nameStrings = NULL)
}
\arguments{
\item{x}{the vector (or scalar) to be named}

\item{decreasing}{if naming is required, how should \code{x} be sorted?}

\item{nameStrings}{an optional vector that provides the names to be used.
See Usage Notes below}
}
\value{
If \code{x} is \code{NULL} or named, it is returned unaltered.  Otherwise,
it is first sorted according to \code{descending} and then names from \code{names} are
assigned in order.
}
\description{
Ensures that a vector or scalar is named according to standard rules
}
\section{Usage Notes}{

If \code{nameStrings} is not \code{NULL}, its length must be greater than or equal to
the length of \code{x}.  Otherwise, names are determined as follows:
If \code{length(x)} is
\code{1}, the name of \code{nameString}'s single element is \code{"action"}
\code{2}, the names of \code{nameString}'s elements are \code{"action"} and \code{"warn"}, in
that order
\code{2} or more, \code{nameString}'s elements are the character equivalent of the
integers \code{1} to \code{length(x)}.
Otherwise t
}

