% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl-evaluation-methods.R
\name{evaluatePointEstimateQTL}
\alias{evaluatePointEstimateQTL}
\title{Compares a scalar statistic derived from the posterior with one or more fixed values}
\usage{
evaluatePointEstimateQTL(
  data,
  posterior,
  metric,
  stat = mean,
  observedMetric,
  statusCol = Status,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{data}{the tibble containing site-level observed metrics (KRIs)}

\item{posterior}{the tibble containing the posterior distribution of the}

\item{metric}{the column in \code{posterior} on which the metric should be based}

\item{stat}{the summary statistic to be calculated from the values in
\code{metric}.  Default: \code{mean}.}

\item{observedMetric}{the column in \code{data} containing the site level
observed metrics (KRIs)}

\item{statusCol}{the name of the column to be added to \code{data} that will
contain the site-level flags comparing the corresponding KRI to the QTL
thresholds defined by \code{lower} and \code{upper}}

\item{lower}{a scalar or vector of lower limits, or NULL.  See Usage Notes
below.}

\item{upper}{a scalar or vector of upper limits, or NULL.  See Usage Notes
below.}
}
\value{
a list of three elements named \code{status}, \code{data} and \code{qtl}.  If
\code{status} is \code{"OK"}, no breach of the QTL was detected.  Otherwise, \code{status}
is equal to name of the most serous breach of the QTL as determined by the
values in either \code{lower} or \code{upper}.
}
\description{
Compares a scalar statistic derived from the posterior with one or more fixed values
}
\section{Usage Notes}{

Both \code{lower} and \code{upper} cannot be NULL.  In what follows, \code{limit} refers to
both \code{lower} and \code{upper}.

If \code{limit} has names, those names are used.  Otherwise, names are assigned
based on the length of \code{limit}.  If \code{length(limit)} is \code{1}, the single
element is named \code{"action"}.  If \code{length(limit)} is \code{2}, the two elements are
named \code{"warn"} and \code{"action"}.  (In this case, \code{limit} is first sorted
appropriately.)  If \code{length(limit)} is \verb{>2}, its elements are named \code{"1"} to
\code{"n"} where \code{n} is \code{length(limit)}.
}

\examples{
post <- (berrySummary \%>\% fitBayesBinomialModel(Subjects, Events))$tab
}
