% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl-evaluation-methods.R
\name{evaluateProbabilityInRangeQTL}
\alias{evaluateProbabilityInRangeQTL}
\title{Evaluates a QTL based on prob(study-level metric lies within a range)}
\usage{
evaluateProbabilityInRangeQTL(
  data,
  posterior,
  metric,
  observedMetric,
  statusCol = Status,
  range,
  probs,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{data}{the tibble containing site-level observed metrics (KRIs)}

\item{posterior}{the tibble containing the posterior distribution of the,
metric, usually obtained from a fit Bayes model function.}

\item{metric}{the column in \code{posterior} on which the QTL should be based}

\item{observedMetric}{the column in \code{data} containing the site level
observed metrics (KRIs)}

\item{statusCol}{the name of the column to be added to \code{data} that will
contain the site-level flags comparing the corresponding KRI to the QTL
thresholds defined by \code{lower} and \code{upper}}

\item{range}{the range of values of \code{metric} to be used when calculating
\code{qtl} = Prob(\code{metric} in \code{range}).}

\item{probs}{the thresholds for Prob(\code{metric} in \code{range}) that define the
action and warning limits for the QTL}

\item{lower}{a scalar or vector of lower limits, or NULL.  See Usage Notes
below.}

\item{upper}{a scalar or vector of upper limits, or NULL.  See Usage Notes
below.}
}
\value{
a list of three elements named \code{status}, \code{data} and \code{qtl}.  If
\code{status} is \code{"OK"}, no breach of the QTL was detected.  Otherwise, \code{status}
is equal to name of the most serious breach of the QTL as determined by the
values in either \code{lower} or \code{upper}. \code{qtl} contains the posterior probability
that \code{metric} lies within \code{range}.
}
\description{
Evaluates the probability that a study-level metric lies within a given range
based in the posterior density provided by a call to a fit Bayes XXXX Model
function.  Action and warning limits are determined by parameters \code{range} and
\code{probs}.  That is, for a single range for the metric, various limits are
defined, each based on an element of \code{probs}.  Site-level KRIs are flagged
based on the values in \code{lower} and \code{upper}.
}
\details{
For more details, see Usage Notes below.
}
\section{Usage Notes}{

Either or both \code{lower} of \code{upper} can be NULL.  In what follows, \code{limit} refers to
both \code{lower} and \code{upper}.
#' If \code{limit} has names, those names are used.  Otherwise, names are assigned
based on the length of \code{limit}.  If \code{length(limit)} is \code{1}, the single
element is named \code{"action"}.  If \code{length(limit)} is \code{2}, the two elements are
named \code{"warn"} and \code{"action"}.  (In this case, \code{limit} is first sorted
appropriately.)  If \code{length(limit)} is \verb{>2}, its elements are named \code{"1"} to
\code{"n"} where \code{n} is \code{length(limit)}.
}

