% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl-evaluation-methods.R
\name{evaluateSiteMetricQTL}
\alias{evaluateSiteMetricQTL}
\title{Evaluates a QTL based on the proportion of site level KRIs within a range}
\usage{
evaluateSiteMetricQTL(
  data,
  posterior,
  metric,
  observedMetric,
  statusCol = Status,
  statusFunc = NULL,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{data}{the tibble containing site-level observed metrics (KRIs)}

\item{posterior}{the tibble containing the posterior distribution of the,
metric, usually obtained from a fit Bayes model function.}

\item{metric}{the column in \code{posterior} on which the QTL should be based}

\item{observedMetric}{the column in \code{data} containing the site level
observed metrics (KRIs)}

\item{statusCol}{the name of the column to be added to \code{data} that will
contain the site-level flags comparing the corresponding KRI to the QTL
thresholds defined by \code{lower} and \code{upper}}

\item{statusFunc}{a function that takes as its single argument the \code{qtl}
element of the return value and converts it to a status string}

\item{lower}{a scalar or vector of lower limits, or NULL.  See Usage Notes
below.}

\item{upper}{a scalar or vector of upper limits, or NULL.  See Usage Notes
below.}

\item{range}{the range of values of \code{metric} to be used when calculating
\code{qtl} = Prob(\code{metric} in \code{range}).}

\item{probs}{the thresholds for Prob(\code{metric} in \code{range}) that define the
action and warning limits for the QTL}
}
\value{
a list of three elements named \code{status}, \code{data} and \code{qtl}.  If
\code{status} is \code{"OK"}, no breach of the QTL was detected.  Otherwise, \code{status}
is equal to name of the most serious breach of the QTL as determined by the
values in either \code{lower} or \code{upper}. \code{qtl} is a tibble whose rows provide
counts of the number of sites whose KRIs lie in the ranges defined by \code{lower}
and \code{upper}.  In addition, the \code{quantiles} element provides the conversion
between quantiles of the posterior distribution of \code{metric} and values of
\code{metric} itself.
}
\description{
Calculates the proportion of site-level KRIs that lie within a range
defined by quantiles of the posterior density provided by a call to a
fit Bayes XXXX Model function.
}
\details{
For more details, see Usage Notes below.
}
\section{Usage Notes}{

Either or both \code{lower} of \code{upper} can be NULL.  In what follows, \code{limit} refers to
both \code{lower} and \code{upper}.
#' If \code{limit} has names, those names are used.  Otherwise, names are assigned
based on the length of \code{limit}.  If \code{length(limit)} is \code{1}, the single
element is named \code{"action"}.  If \code{length(limit)} is \code{2}, the two elements are
named \code{"warn"} and \code{"action"}.  (In this case, \code{limit} is first sorted
appropriately.)  If \code{length(limit)} is \verb{>2}, its elements are named \code{"1"} to
\code{"n"} where \code{n} is \code{length(limit)}.

If \code{statusFunc} is \code{NULL}, then the \code{status} element of the return value is
set to \code{"OK"} by default.
}

\examples{
berrySummary \%>\%
evaluateSiteMetricQTL(
  posterior=fitted$tab,
  metric=p,
  observedMetric=ObservedResponse,
  lower=c("action"=0.5, "warn"=0.6),
  upper=c("action"=0.9, "warn"=0.8),
  statusFunc=function(d) ifelse(
                           d \%>\% 
                             dplyr::filter(Status == "action") \%>\% 
                             dplyr::pull(N) > 1, "action", "OK"
                         )
)
}
