% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelString.R
\name{getModelString}
\alias{getModelString}
\title{Function to obtain the string that defines the default JAGS model for each
data type}
\usage{
getModelString(
  label = c("binary", "binomial", "normal", "poisson", "tte"),
  hyperParams = list(binomial = c(`<priorA>` = "dunif(0, 10)", `<priorB>` =
    "dunif(0, 10)"), binary = c(`<priorA>` = "dunif(0, 10)", `<priorB>` =
    "dunif(0, 10)"), normal = c(`<priorMu0>` = "dnorm(0, 1e-06)", `<priorInvTau>` =
    "dgamma(1e-06, 1e-06)"), poisson = c(`<priorShape>` = "dgamma(1, 1)", `<priorScale>`
    = "dgamma(1, 1)"), tte = c(`<priorMu0>` = "dnorm(0, 1e-06)", `<priorInvTau>` =
    "dgamma(1e-06, 1e-06)")),
  prior = FALSE
)
}
\arguments{
\item{label}{the type of model}

\item{hyperParams}{a named list or named vector that defines how the
placeholders in the model string are to be handled.  See Usage Notes below.}

\item{prior}{Boolean.  Should the model for the prior be returned.  Default:
\code{FALSE}.}
}
\value{
a string defining the required model
}
\description{
Function to obtain the string that defines the default JAGS model for each
data type
}
\section{Usage Notes}{

If \code{hyperParams} is a list, then the placeholder replacements are taken from
\code{hyperParams[[label]]}, which should be a named vector.  Otherwise, the
placeholder replacements are taken from \code{hyperParams} directly.  Again,
\code{hyperParams} should be a named vector.  Once a named vector has been
obtained from \code{hyperParams}, names of \code{hyperParams} define the placeholders
and the replacements by the corresponding values.

For example, \code{hyperParams} has the value \verb{c("<priorA>"="dunif(0, 10)",}
\verb{"<priorB>"="dunif(0, 20)")}, then where ever \verb{<priorA> is found in the model string, it is replaced by }dunif(0, 10)\verb{.  Similarly, all occurrences of }\if{html}{\out{<priorB>}}\verb{will be replaced by}dunif(0, 20)`.

The default value of \code{hyperParams} provides the default hyperparameter values
used by \code{rbqmR}'s Bayes model fitting functions.
}

\section{Usage notes}{

\code{label} is determined case insensitively
}

\examples{
#To fit a binary model
binModel <- getModelString("binary")
}
