% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteRates}
\alias{siteRates}
\title{A dataset of event rates}
\format{
A tibble with 47 rows and 6 columns
\describe{
\item{Site}{The site ID}
\item{Country}{The subject ID}
\item{Patients}{The number of patients at the site}
\item{Exposure}{The total exposure at the site, in months}
\item{Events}{The total number of events at the site}
\item{Rate}{The site's event rate (events per month)}
}
}
\usage{
siteRates
}
\description{
A dataset of event rates
}
\keyword{datasets}
