% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaLung}
\alias{vaLung}
\title{The Kalbfleisch and Prentice (1980) VA lung dataset.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 228 rows and 11 columns.
}
\usage{
vaLung
}
\description{
@format A tibble with 137 rows and 8 variables:
\describe{
\item{Treatment}{Treatment group}
\item{CellType}{cell type}
\item{SurvivalTime}{Survival time sice randomisation}
\item{Status}{1=Event, 0=Censored}
\item{KPS}{Karnofsky Performance Status}
\item{Age}{Age in years at randomisation}
\item{PriorTherapy}{Boolean.  received prior therapy?}
}
@source Kalbfleisch D and Prentice RL (1980), The Statistical Analysis of
Failure Time Data. Wiley, New York
}
\keyword{datasets}
