% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQtlPlot.R
\name{createQtlPlot}
\alias{createQtlPlot}
\title{Summary Plot of Observed Event Rates/Proportions}
\usage{
createQtlPlot(
  data,
  metric = p,
  siteData = NULL,
  targetRange = NULL,
  actionLimits = NULL,
  warningLimits = NULL,
  observedMetric = NULL,
  siteSize = NULL,
  siteMetric = NULL
)
}
\arguments{
\item{data}{A tibble containing MCMC samples for which the density is
required}

\item{metric}{The column in data that contains the sampled values of the metric
being evaluated}

\item{siteData}{A tibble containing the observed site specific data}

\item{targetRange}{A tibble containing the lower and upper limits of the
target range for the metric}

\item{actionLimits}{a list or named vector defining the action limits that
will appear on the graph. See \verb{[shadeRange]} for details.  May be \code{NA} or
\code{NULL}.}

\item{warningLimits}{a list or named vector defining the action limits that
will appear on the graph. See \verb{[shadeRange]} for details.  May be \code{NA} or
\code{NULL}.}

\item{observedMetric}{the observed value of the metric}

\item{siteSize}{The column in \code{siteData} that defines the sample size at each
site.  Uses tidyselect. Ignored if \code{siteData} is \code{NULL}}

\item{siteMetric}{The column in \code{siteData} that defines the KRI at the site.
Uses tidyselect. Ignored if \code{siteData} is \code{NULL}.}
}
\description{
Summary Plot of Observed Event Rates/Proportions
}
