% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeRange.R
\name{shadeRange}
\alias{shadeRange}
\title{Shade Areas Under The Curve}
\usage{
shadeRange(
  plot,
  range = list(list(lower = NA, upper = 0.1, alpha = 0.3, colour = "goldenrod1"),
    list(lower = 0.9, upper = NA, alpha = 0.3, colour = "goldenrod1")),
  idx = 2
)
}
\arguments{
\item{plot}{The \code{ggplot2} object}

\item{range}{a \code{vector} (or \verb{list of vectors}) containing the definition(s) of
the range(s) to be shaded.  See Usage Notes below.}

\item{idx}{An integer defining the index within the list returned by
\code{ggplot2::ggpot_build} that identifies the layer containing the density to be
shaded.  Typicall, this corresponds to the order in which the various \code{geom}s
are added to the \code{ggplot2} object.  For example, in \code{d \%>\% ggplot2() \%>\%  geom_point(...) \%>\% geom_density(...)}, \verb{idx should be }2`.}
}
\value{
The modified \code{ggplot2} object
}
\description{
Given an existing \code{ggplot2} object, shade an area under an existing density
curve.
}
\section{Usage Notes}{

\code{range} should be a list or a list of vectors.  If a list, it should contain
named elements \code{lower}, \code{upper}, \code{alpha} and \code{colour}, where \code{lower} and
\code{upper} contain the lower and upper x axis values that define the extent of
the shaded area, \code{colour} defines the colour of the shaded area and \code{alpha}
defines the alpha value that defines the transparency of the shading.  An
\code{alpha} of \code{0} denotes complete transparency (or invisibility).  An \code{alpha}
of 1 denotes total opacity, meaning that the contents of any layers below
the density will be invisible if within the shaded area.

To define shaded ranges that include the left- or right-most extents of the
density, set \code{lower} or \code{upper} respectively to \code{NA}.

If a list of lists, each element of the  list should be a list with
elements as defined above.

For example, the default value of \code{range} defines the ranges to be shaded as
those that lie below the 10th centile or above the 90th centile of the
density in goldenrod1 using an alpha of 0.3.
}

